+++ src/test/standalone/build.test.ts
@@ -1,9 +1,7 @@
 "use strict"
 import * as assert from "assert"
 import * as fs from "fs"
-import MarkdownIt from "markdown-it"
 import * as path from "path"
-import { mystBlockPlugin } from "../../mdPlugins"
 import { readGrammar, readSnippets } from "../../readTemplates"
 
 function sanitize(fileString: string) {
@@ -30,21 +28,4 @@ suite("Build Tests", () => {
     const jsonString = readSnippets(true)
     assert.equal(sanitize(jsonString as string), sanitize(expected))
   })
-})
-
-suite("Syntax Fixtures: MyST Blocks", () => {
-  const fixtures = fs.readFileSync(
-    path.join(__dirname, "../../../test_static/syntax-fixtures", "myst_block.md"),
-    "utf8"
-  )
-  const mdit = MarkdownIt("commonmark").use(mystBlockPlugin)
-  fixtures
-    .split("\n.\n\n")
-    .map(s => s.split("\n.\n"))
-    .forEach(([name, text, expected]) => {
-      test(name, () => {
-        const rendered = mdit.render(text)
-        assert.equal(rendered.trim(), expected.trim())
-      })
-    })
 })
